+++ client/store/index.js
@@ -10,7 +10,6 @@ import categories from './categories'
 import activeSizes from './active-sizes'
 import reviews from './reviews'
 import searchTerm from './search-term'
-import orderStatus from './order-status'
 import orders from './orders'
 
 const reducer = combineReducers({
@@ -22,7 +21,6 @@ const reducer = combineReducers({
   activeSizes,
   searchTerm,
   reviews,
-  orderStatus,
   orders
 })
 const middleware = composeWithDevTools(applyMiddleware(
@@ -40,5 +38,4 @@ export * from './categories'
 export * from './active-sizes'
 export * from './reviews'
 export * from './search-term'
-export * from './order-status'
 export * from './orders'
