+++ client/store/orders.js
@@ -1,9 +1,11 @@
 import axios from 'axios'
+import history from '../history'
 
 /**
  * ACTION TYPES
  */
 const GOT_ORDERS = 'GOT_ORDERS'
+const CHANGE_ORDER_STATUS = 'CHANGE_ORDER_STATUS'
 
 /**
  * INITIAL STATE
@@ -16,6 +18,11 @@ const defaultOrders = []
 export const gotOrders = orders => ({
   type: GOT_ORDERS,
   orders
+})
+
+const changeOrderStatus = order => ({
+  type: CHANGE_ORDER_STATUS,
+  order
 })
 
 /**
@@ -39,13 +46,30 @@ export const fetchOrdersByCustomerId = (customerid) =>
       })
       .catch(err => console.error('fetching orders went wrong', err))
 
+
+export const updateOrderStatus = (id, newStatus) =>
+  dispatch => {
+    return axios.put(`/api/orders/${id}/`, { orderStatus: newStatus })
+      .then(res => res.data)
+      .then(order => {
+        dispatch(changeOrderStatus(order))
+        history.push('/admin/orders/')
+      })
+      .catch(err => console.error('updating order status went wrong', err))
+
+  }
+
 /**
  * REDUCER
  */
-export default function (state = defaultOrders, action){
+export default function (state = defaultOrders, action) {
   switch (action.type) {
     case GOT_ORDERS:
       return action.orders
+    case CHANGE_ORDER_STATUS: {
+      return state.map(order => (
+        action.order.id === order.id ? action.order : order))
+    }
     default: return state
   }
 }
